requirements.txt
fastapi
uvicorn
python-multipart
pandas
pdfplumber
python-docx
langchain
langchain-community
sentence-transformers
faiss-cpu
groq
langchain-groq
